tests: refactor folder structure;
