+++ tests/type_hints/test_contract_type_hint.py
